# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Swagger Project","description":"Swagger Documentaion for Project","termsOfService":"http:\/\/swagger.io\/terms\/","contact":{"email":"gaurav.bakshi@studiographene.com"},"version":"1.0.0"},"paths":{"\/api\/login":{"post":{"tags":["Login"],"summary":"To login to application","description":"Method to login via API","operationId":"login","parameters":[{"name":"project-device-id","in":"header","description":"This is unique device identifier.","required":true,"schema":{"type":"string"}},{"name":"project-device-type","in":"header","description":"This is device type.e.g: ios","required":true,"schema":{"type":"string"}},{"name":"Accept","in":"header","description":"application\/json,","required":true,"schema":{"type":"string"}},{"name":"Content-Type","in":"header","description":"Content-Type","required":true,"schema":{"type":"string"}}],"requestBody":{"required":false,"content":{"application\/x-www-form-urlencoded":{"schema":{"properties":{"email":{"description":"User email","type":"string"},"password":{"description":"User password","type":"string"}},"type":"object"}}}},"responses":{"200":{"description":"User details with token."},"default":{"description":"Error message with details."},"400":{"description":"Invalid username\/password supplied."},"500":{"description":"Application issue, Please contact API maintenance Team."}},"deprecated":false}},"\/api\/logout":{"get":{"tags":["Logout"],"summary":"To logout of application","description":"Method to logout via API","parameters":[{"name":"project-device-id","in":"header","description":"This is unique device identifier.","required":true,"schema":{"type":"string"}},{"name":"project-device-type","in":"header","description":"This is device type.e.g: ios","required":true,"schema":{"type":"string"}},{"name":"Accept","in":"header","description":"application\/json","required":true,"schema":{"type":"string"}},{"name":"Content-Type","in":"header","description":"Content-Type","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User details with token."},"default":{"description":"Error message with details"},"400":{"description":"Invalid token supplied."},"500":{"description":"Application issue, Please contact API maintenance Team."}},"security":[{"api_auth":[]}]}}},"components":{"securitySchemes":{"api_auth":{"type":"apiKey","description":"A short description for security scheme","name":"Authorization","in":"header"}}}}